+++ frameworks/react/components/forms/radio-group/RadioGroup.d.ts
@@ -2,6 +2,8 @@ import * as React from 'react';
 
 export interface RadioGroupProps {
 
+  ariaLabel: string;
+
   children?: React.ReactNode;
 
   value?: string;